build_dir.pl - a script to build a directory stucture for a project in a common way. This means that EMR bioinformatics packages can input and export files to consistent locations.
idris.csv - an example of an input .csv file that build_dir.pl can use to constuct a directory structure. The entries in this file should be changed for each new project. The filename should be the name of the project as this will be used to name the project directory.